+++ src/components/EtherScanTxLink.tsx
@@ -1,22 +1,30 @@
+import etherscanLink from "@metamask/etherscan-link";
+
 interface TxHashLinkProps {
   txHash: string;
+  providerChainId: bigint | undefined;
 }
 
 /**
  * A component that displays the EtherScan link of a transaction hash.
  */
 export const TxHashLink = function TxHashLink(props: TxHashLinkProps) {
-  const { txHash } = props;
+  const { txHash, providerChainId } = props;
+
+  let txLink;
+  try {
+    txLink = etherscanLink.createExplorerLink(
+      txHash,
+      providerChainId?.toString() ?? "-1",
+    );
+  } catch (e) {
+    console.error(e);
+  }
 
   return (
     <div className="centered">
-      Txhash: {/* TODO: adjust link for different networks */}
-      <a
-        className="link"
-        target="_blank"
-        rel="noreferrer"
-        href={`https://sepolia.etherscan.io/tx/${txHash}`}
-      >
+      Txhash:
+      <a className="link" target="_blank" rel="noreferrer" href={`${txLink}`}>
         {txHash}
       </a>
     </div>