+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 ## Unreleased changes
 
+- Add handling of "Injected Network (by wallet)".
 - Add handling of different blockchain networks (`Ethereum` and `Sepolia`).
 - Add textfield option to enable deriving the function interface from raw `byteCode`.
 - Add decoding of return parameter types.
 
 ## 1.0.0
 
-- Initial `front end` to interact with any contract (no ABI/source coded needed).
+- Initial `front end` to interact with any contract (no ABI/source code needed).